requirements.txt
google-cloud-aiplatform >= 1.31.0
flask
google-cloud-logging
google-genai 
langchain-google-vertexai 
langchain